pyproject.toml
[tool.poetry]
name = "personalcli"
version = "0.1.0"
description = ""
authors = ["Rafael Garcia <rafa.informatica@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
personalcli = "personalcli.main:app"

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.12.3"}
pydantic = "^2.7.4"
pydantic-settings = "^2.3.3"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-clarity = "^1.0.1"
pytest-testmon = "^2.1.1"
pytest-watch = "^4.2.0"


[tool.poetry.group.gemini.dependencies]
google-generativeai = "^0.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
